==> 2_reverse_geocoder.ipynb <==
import numpy as np
import pandas as pd

import json
import requests
import sys

from google.colab import drive 
drive.mount('/content/gdrive')

df = pd.read_csv('gdrive/My Drive/Colab Notebooks/businesses_capstone/chicago_new.csv')

df.info()

# 54,446 unique latlong combinations
len(df['latlong'].unique())

df_latlong = pd.DataFrame(df['latlong'].unique(), columns = ['latlong'])
df_latlong

# Test
df_test= df.iloc[0:5]

df_test.info()

# Test - Google Reverse Geocoder API

data = []
for index, ltlng in enumerate(np.unique(df_test['latlong'].values)):
    try:
        response = requests.get(
            'https://maps.googleapis.com/maps/api/geocode/json',
            params={
            'latlng': ltlng,
            'key':"AIzaSyAnQB5SvuzcaCMzuLBH9vZTSfvXsStDoo4" },
            )
        types=response.json()['results'][0]['types'][0]
        for i in response.json()['results'][0]['address_components']:
            if i['types'][0]=='street_number':
                street_number=(i['long_name'])
            if i['types'][0]=='route':
                route=(i['long_name'])
            if i['types'][0]=='neighborhood':
                neighborhood=(i['long_name'])
            if i['types'][0]=='locality':
                locality =(i['long_name'])
            if i['types'][0]=='administrative_area_level_1':
                state2 =(i['short_name'])
        data.append({
                'index':index,
                'key':ltlng,
                'types':types,
                'street_number':street_number,
                'route':route,
                'neighborhood':neighborhood,
                'locality':locality,
                'state2':state2
                })
        if index%500==0:
            print(index)
    except Exception as e:
            print(e)
            exc_type, exc_obj, exc_tb = sys.exc_info()
            print(idx, exc_type, ltlng, exc_tb.tb_lineno)

# Code source: https://github.com/brandonanhorn/Pokemon_GO_ML/blob/master/Pokemon-Go/Reversegeocoder.ipynb

data

df_test_final = pd.DataFrame(data, columns = ['index','key', 'types', 'street_number', 'route', 'neighborhood', 'locality', 'state2'])

type(df_test_final)

df_test_final

df_test_final.columns

# Google Reverse Geocoder API - all unique longlat values

data = []
for index, ltlng in enumerate(np.unique(df_latlong['latlong'].values)):
    try:
        response = requests.get(
            'https://maps.googleapis.com/maps/api/geocode/json',
            params={
            'latlng': ltlng,
            'key':"AIzaSyAnQB5SvuzcaCMzuLBH9vZTSfvXsStDoo4" },
            )
        types=response.json()['results'][0]['types'][0]
        for i in response.json()['results'][0]['address_components']:
            if i['types'][0]=='street_number':
                street_number=(i['long_name'])
            if i['types'][0]=='route':
                route=(i['long_name'])
            if i['types'][0]=='neighborhood':
                neighborhood=(i['long_name'])
            if i['types'][0]=='locality':
                locality =(i['long_name'])
            if i['types'][0]=='administrative_area_level_1':
                state2 =(i['short_name'])
        data.append({
                'index':index,
                'key':ltlng,
                'types':types,
                'street_number':street_number,
                'route':route,
                'neighborhood':neighborhood,
                'locality':locality,
                'state2':state2
                })
        if index%500==0:
            print(index)
    except Exception as e:
            print(e)
            exc_type, exc_obj, exc_tb = sys.exc_info()
            print(idx, exc_type, ltlng, exc_tb.tb_lineno)

# Code source: https://github.com/brandonanhorn/Pokemon_GO_ML/blob/master/Pokemon-Go/Reversegeocoder.ipynb

data[5:10]

len(data)

df_neighborhoods = pd.DataFrame(data, columns = ['index','key', 'types', 'street_number', 'route', 'neighborhood', 'locality', 'state2'])

df_neighborhoods.head()

df_neighborhoods.columns

df_neighborhoods.to_csv('neighborhood_results.csv', header=True, index=False)

chicago_final = pd.merge(left=df, right=df_neighborhoods, how='left', left_on='latlong', right_on='key')

chicago_final.info()

chicago_final.to_csv('chicago_final.csv', header=True, index=False)

chicago_final.columns

chicago_final.head().T
